+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+UNRELEASED -
+-------------------
+
+## Fixed
+- Improved resiliency of the default minion config for a saltmaster failover.
+
+
 0.16.1 - 2023-10-17
 -------------------